move compute strides into KokkosFFT_Extents.hpp
