+++ common/src/KokkosFFT_Extents.hpp
@@ -48,6 +48,43 @@ auto extract_extents(const ViewType& view) {
   return extents;
 }
 
+// \brief Helper to compute strides from extents.
+// The extents are computed from a LayoutRight View.
+// The computed strides can be considered as view strides
+// in the reversed order.
+//
+// Examples:
+// v0 (n0) -> (v0.stride(0)) or (1)
+// v1 (n0, n1) -> (v1.stride(1), v1.stride(0)) or (1, n1)
+// v2 (n0, n1, n2) -> (v2.stride(2), v2.stride(1), v2.stride(0))
+//                 or (1, n2, n2 * n1)
+/// \tparam ContainerType The container type, must be either one of std::array
+/// or std::vector
+/// \param[in] extents The extents of the data
+/// \return strides computed from the input data
+/// \throws runtime_error if any of the extents is less than or equal to 0
+template <typename ContainerType,
+          std::enable_if_t<is_std_vector_v<ContainerType> ||
+                               is_std_array_v<ContainerType>,
+                           std::nullptr_t> = nullptr>
+auto compute_strides(const ContainerType& extents) {
+  using index_type = std::remove_cv_t<
+      std::remove_reference_t<typename ContainerType::value_type>>;
+  static_assert(std::is_integral_v<index_type>,
+                "compute_strides: index_type must be an integral type.");
+  KOKKOSFFT_THROW_IF(std::any_of(extents.begin(), extents.end(),
+                                 [](index_type extent) { return extent <= 0; }),
+                     "extents must be greater than 0");
+  ContainerType strides = extents, reversed_extents = extents;
+  std::reverse(reversed_extents.begin(), reversed_extents.end());
+
+  strides.at(0) = 1;
+  for (std::size_t i = 1; i < reversed_extents.size(); i++) {
+    strides.at(i) = reversed_extents.at(i - 1) * strides.at(i - 1);
+  }
+  return strides;
+}
+
 /// \brief Return a new shape of the input view based on the
 /// specified input shape and axes.
 ///
