+++ common/src/KokkosFFT_utils.hpp
@@ -331,42 +331,6 @@ auto convert_base_int_type(const ContainerType& src) {
   }
 }
 
-// \brief Helper to compute strides from extents.
-// The extents are computed from a LayoutRight View.
-// The computed strides can be considered as view strides
-// in the reversed order.
-//
-// Examples:
-// v0 (n0) -> (v0.stride(0)) or (1)
-// v1 (n0, n1) -> (v1.stride(1), v1.stride(0)) or (1, n1)
-// v2 (n0, n1, n2) -> (v2.stride(2), v2.stride(1), v2.stride(0))
-//                 or (1, n2, n2 * n1)
-/// \tparam ContainerType The container type, must be either one of std::array
-/// or std::vector
-/// \param[in] extents The extents of the data
-/// \return strides computed from the input data
-template <typename ContainerType,
-          std::enable_if_t<is_std_vector_v<ContainerType> ||
-                               is_std_array_v<ContainerType>,
-                           std::nullptr_t> = nullptr>
-auto compute_strides(const ContainerType& extents) {
-  using index_type =
-      std::remove_cv_t<std::remove_reference_t<decltype(*extents.begin())>>;
-  static_assert(std::is_integral_v<index_type>,
-                "compute_strides: index_type must be an integral type.");
-  KOKKOSFFT_THROW_IF(
-      total_size(extents) <= 0,
-      "compute_strides: total size of the extents must not be 0");
-  ContainerType strides = extents, reversed_extents = extents;
-  std::reverse(reversed_extents.begin(), reversed_extents.end());
-
-  strides.at(0) = 1;
-  for (std::size_t i = 1; i < reversed_extents.size(); i++) {
-    strides.at(i) = reversed_extents.at(i - 1) * strides.at(i - 1);
-  }
-  return strides;
-}
-
 /// \brief Returns a reversed copy of the input container
 /// \tparam Container The container type, must support begin() and end()
 /// \param[in,out] c The input container
